LRCI-3892 Remove unnecessary JobBuilds component
