--- workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-custom-element/src/components/JobBuilds/JobBuilds.js
@@ -1,102 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import ClayPanel from '@clayui/panel';
-import {useState} from 'react';
-import {Link} from 'react-router-dom';
-
-import {toLocaleString} from '../../services/DateUtil';
-import {toDurationString} from '../../services/DurationUtil';
-import useSpringBootData from '../../services/useSpringBootData';
-import Jethr0Table from '../Jethr0Table/Jethr0Table';
-
-function JobBuilds({jobId}) {
-	const [jobBuilds, setJobBuilds] = useState(null);
-
-	useSpringBootData({
-		setData: setJobBuilds,
-		urlPath: '/jobs/builds/' + jobId,
-	});
-
-	if (!jobBuilds) {
-		return <div>Loading...</div>;
-	}
-
-	return (
-		<ClayPanel
-			collapsable
-			defaultExpanded
-			displayTitle="Builds"
-			displayType="secondary"
-			showCollapseIcon={true}
-		>
-			<ClayPanel.Body>
-				<Jethr0Table>
-					<thead>
-						<tr>
-							<th>ID</th>
-							<th>Name</th>
-							<th>Create Date</th>
-							<th>State</th>
-							<th>Initial Build</th>
-							<th>Jenkins Duration</th>
-							<th>Jenkins Build</th>
-						</tr>
-					</thead>
-					<tbody>
-						{jobBuilds &&
-							jobBuilds.map((jobBuild) => {
-								return (
-									<tr key={jobBuild.id}>
-										<th className="font-weight-semi-bold">
-											<Link
-												title={jobBuild.id}
-												to={
-													'/jobs/builds/' +
-													jobBuild.id
-												}
-											>
-												{jobBuild.id}
-											</Link>
-										</th>
-										<td>{jobBuild.name}</td>
-										<td>
-											{toLocaleString(
-												jobBuild.dateCreated
-											)}
-										</td>
-										<td>{jobBuild.state.name}</td>
-										<td>
-											{jobBuild.initialBuild.toString()}
-										</td>
-										<td>
-											{toDurationString(
-												jobBuild.latestDuration
-											)}
-										</td>
-										<td>
-											{jobBuild.latestJenkinsBuildURL ? (
-												<a
-													href={
-														jobBuild.latestJenkinsBuildURL
-													}
-												>
-													Latest Jenkins Build
-												</a>
-											) : (
-												<div>-</div>
-											)}
-										</td>
-									</tr>
-								);
-							})}
-					</tbody>
-				</Jethr0Table>
-			</ClayPanel.Body>
-		</ClayPanel>
-	);
-}
-
-export default JobBuilds;
